♻️ 8EM383 upgrade: decompose apply git helpers
