+++ .agentplane/tasks/202605290426-8EM383/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ .../agentplane/src/commands/upgrade/apply-git.ts   | 294 +++++++++++++++++++++
+ packages/agentplane/src/commands/upgrade/apply.ts  | 289 +-------------------
+ 2 files changed, 295 insertions(+), 288 deletions(-)
