+++ .agentplane/tasks/202605290426-8EM383/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 8EM383 task: Upgrade apply command decomposition [202605290426-8EM383]
