+++ .agentplane/tasks/202605290426-8EM383/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290426-8EM383/upgrade-apply-command-decomposition",
+  "created_at": "2026-05-29T04:27:13.087Z",
+  "diffstat_sha256": "sha256:4326fd044b765542cfa4551bf30529dcc08b2a71c2135d0ffe323535bad4dcf4",
+  "last_verified_at": "2026-05-29T04:33:19.669Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290426-8EM383",
+  "updated_at": "2026-05-29T04:27:13.087Z",
+  "verify": {
+    "status": "pass"
+  }
+}
